requirements/prod.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements/prod.in
#
ansi2html==1.6.0
    # via -r requirements/prod.in
appdirs==1.4.4
    # via
    #   cumulusci
    #   fs
asgiref==3.5.0
    # via django
authlib==0.15.5
    # via
    #   cumulusci
    #   simple-salesforce
bleach==4.1.0
    # via sfdo-template-helpers
boto3==1.20.42
    # via -r requirements/prod.in
botocore==1.23.42
    # via
    #   boto3
    #   s3transfer
certifi==2021.10.8
    # via
    #   cumulusci
    #   requests
    #   snowfakery
cffi==1.15.0
    # via
    #   cryptography
    #   cumulusci
charset-normalizer==2.0.10
    # via
    #   cumulusci
    #   requests
    #   snowfakery
click==7.1.2
    # via
    #   cumulusci
    #   rq
    #   snowfakery
colorama==0.4.4
    # via
    #   cumulusci
    #   rich
coloredlogs==15.0.1
    # via -r requirements/prod.in
commonmark==0.9.1
    # via
    #   cumulusci
    #   rich
croniter==1.2.0
    # via rq-scheduler
cryptography==36.0.1
    # via
    #   authlib
    #   cumulusci
    #   jwcrypto
    #   pyjwt
    #   secretstorage
    #   sfdo-template-helpers
cumulusci==3.51.1
    # via -r requirements/prod.in
defusedxml==0.7.1
    # via python3-openid
deprecated==1.2.13
    # via
    #   cumulusci
    #   jwcrypto
    #   redis
django==3.1.14
    # via
    #   -r requirements/prod.in
    #   django-allauth
    #   django-anymail
    #   django-defender
    #   django-filter
    #   django-guardian
    #   django-js-reverse
    #   django-log-request-id
    #   django-model-utils
    #   django-redis
    #   django-rq
    #   django-storages
    #   djangorestframework
    #   sfdo-template-helpers
django-allauth==0.47.0
    # via -r requirements/prod.in
django-anymail==8.5
    # via -r requirements/prod.in
django-crispy-forms==1.14.0
    # via -r requirements/prod.in
django-defender==0.9.1
    # via -r requirements/prod.in
django-environ==0.8.1
    # via -r requirements/prod.in
django-filter==21.1
    # via
    #   -r requirements/prod.in
    #   djangorestframework-filters
    #   sfdo-template-helpers
django-guardian==2.4.0
    # via -r requirements/prod.in
django-js-reverse==0.9.1
    # via -r requirements/prod.in
django-log-request-id==2.0.0
    # via -r requirements/prod.in
django-model-utils==4.2.0
    # via -r requirements/prod.in
django-redis==5.2.0
    # via -r requirements/prod.in
django-rq==2.5.1
    # via -r requirements/prod.in
django-slds-crispyforms==0.1.0
    # via -r requirements/prod.in
django-storages==1.12.3
    # via -r requirements/prod.in
django-watson==1.6.0
    # via -r requirements/prod.in
djangorestframework==3.13.1
    # via
    #   -r requirements/prod.in
    #   djangorestframework-filters
    #   sfdo-template-helpers
djangorestframework-filters==1.0.0.dev2
    # via -r requirements/prod.in
docutils==0.16
    # via cumulusci
faker==11.1.0
    # via
    #   cumulusci
    #   snowfakery
fs==2.4.14
    # via cumulusci
github3.py==3.0.0
    # via cumulusci
greenlet==1.1.2
    # via
    #   cumulusci
    #   snowfakery
    #   sqlalchemy
gunicorn==20.1.0
    # via -r requirements/prod.in
gvgen==1.0
    # via
    #   cumulusci
    #   snowfakery
honcho==1.1.0
    # via -r requirements/prod.in
humanfriendly==10.0
    # via coloredlogs
idna==3.3
    # via
    #   cumulusci
    #   requests
    #   snowfakery
importlib-metadata==4.10.1
    # via
    #   cumulusci
    #   keyring
    #   markdown
jeepney==0.7.1
    # via
    #   keyring
    #   secretstorage
jinja2==2.11.3
    # via
    #   cumulusci
    #   snowfakery
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jwcrypto==1.0
    # via
    #   cumulusci
    #   github3.py
keyring==23.0.1
    # via cumulusci
logfmt==0.4
    # via sfdo-template-helpers
lxml==4.7.1
    # via cumulusci
markdown==3.3.6
    # via sfdo-template-helpers
markupsafe==2.0.1
    # via
    #   cumulusci
    #   jinja2
    #   snowfakery
oauthlib==3.1.1
    # via requests-oauthlib
packaging==21.3
    # via
    #   bleach
    #   redis
psutil==5.9.0
    # via cumulusci
psycopg2-binary==2.9.3
    # via -r requirements/prod.in
pycparser==2.21
    # via
    #   cffi
    #   cumulusci
pydantic==1.9.0
    # via
    #   cumulusci
    #   snowfakery
pygments==2.11.2
    # via
    #   cumulusci
    #   rich
pyjwt[crypto]==2.3.0
    # via
    #   -r requirements/prod.in
    #   cumulusci
    #   django-allauth
pyparsing==3.0.7
    # via packaging
python-baseconv==1.2.2
    # via
    #   cumulusci
    #   snowfakery
python-dateutil==2.8.2
    # via
    #   botocore
    #   croniter
    #   cumulusci
    #   faker
    #   github3.py
    #   rq-scheduler
    #   snowfakery
python3-openid==3.2.0
    # via django-allauth
pytz==2021.3
    # via
    #   cumulusci
    #   django
    #   djangorestframework
    #   fs
pyyaml==6.0
    # via
    #   cumulusci
    #   snowfakery
redis==4.1.1
    # via
    #   -r requirements/prod.in
    #   django-defender
    #   django-redis
    #   django-rq
    #   rq
requests==2.27.0
    # via
    #   cumulusci
    #   django-allauth
    #   django-anymail
    #   github3.py
    #   requests-futures
    #   requests-oauthlib
    #   robotframework-requests
    #   salesforce-bulk
    #   simple-salesforce
    #   snowfakery
requests-futures==1.0.0
    # via cumulusci
requests-oauthlib==1.3.0
    # via django-allauth
rich==11.0.0
    # via cumulusci
robotframework==4.1.3
    # via
    #   cumulusci
    #   robotframework-lint
    #   robotframework-pabot
    #   robotframework-requests
    #   robotframework-seleniumlibrary
    #   robotframework-stacktrace
robotframework-lint==1.1
    # via cumulusci
robotframework-pabot==2.1.0
    # via cumulusci
robotframework-pythonlibcore==3.0.0
    # via
    #   cumulusci
    #   robotframework-seleniumlibrary
robotframework-requests==0.9.2
    # via cumulusci
robotframework-seleniumlibrary==5.1.3
    # via cumulusci
robotframework-stacktrace==0.4.1
    # via
    #   cumulusci
    #   robotframework-pabot
rq==1.10.1
    # via
    #   -r requirements/prod.in
    #   django-rq
    #   rq-scheduler
rq-scheduler==0.11.0
    # via -r requirements/prod.in
rst2ansi==0.1.5
    # via cumulusci
s3transfer==0.5.0
    # via boto3
salesforce-bulk==2.2.0
    # via cumulusci
sarge==0.1.7.post1
    # via cumulusci
secretstorage==3.3.1
    # via keyring
selenium==3.141.0
    # via
    #   cumulusci
    #   robotframework-seleniumlibrary
sfdo-template-helpers @ https://github.com/SFDO-Tooling/sfdo-template-helpers/archive/v0.19.0.tar.gz
    # via -r requirements/prod.in
simple-salesforce==1.11.4
    # via
    #   cumulusci
    #   salesforce-bulk
six==1.16.0
    # via
    #   bleach
    #   cumulusci
    #   fs
    #   python-dateutil
    #   salesforce-bulk
    #   snowfakery
snowfakery==2.4.0
    # via cumulusci
sqlalchemy==1.4.29
    # via
    #   cumulusci
    #   snowfakery
sqlparse==0.4.2
    # via django
text-unidecode==1.3
    # via
    #   cumulusci
    #   faker
    #   snowfakery
typing-extensions==4.0.1
    # via
    #   cumulusci
    #   pydantic
    #   snowfakery
unicodecsv==0.14.1
    # via
    #   cumulusci
    #   salesforce-bulk
uritemplate==4.1.1
    # via
    #   cumulusci
    #   github3.py
urllib3==1.26.7
    # via
    #   botocore
    #   cumulusci
    #   requests
    #   selenium
    #   snowfakery
webencodings==0.5.1
    # via bleach
whitenoise==5.3.0
    # via -r requirements/prod.in
wrapt==1.13.3
    # via
    #   cumulusci
    #   deprecated
xmltodict==0.12.0
    # via cumulusci
zipp==3.7.0
    # via
    #   cumulusci
    #   importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
